pyproject.toml
[tool.poetry]
name = "xbot"
version = "0.1.0"
description = ""
authors = ["Oleh Toporkov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.28"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
openai = "^1.13.3"
pillow = "^10.2.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
